Move socket.io code from main.js server file in to separate module
